Move title and saving to Root component
