+++ static/src/questionnaires/QuestionnairePreview.vue
@@ -1,8 +1,5 @@
 <template>
     <div>
-        <div v-if="hasErrors" class="alert alert-danger">
-          L'envoi de ce formulaire n'a pas fonctionné. Erreur : {{JSON.stringify(errors)}}
-        </div>
         <div class="preview">
             <questionnaire-detail v-bind:questionnaire="questionnaire">
             </questionnaire-detail>
@@ -19,20 +16,12 @@
 </template>
 
 <script>
-    import axios from "axios"
-    import moment from "moment"
     import Vue from "vue"
     import QuestionnaireDetail from "./QuestionnaireDetail"
-
-    const save_questionnaire_url = "/api/questionnaire/"
-    axios.defaults.xsrfCookieName = 'csrftoken'
-    axios.defaults.xsrfHeaderName = 'X-CSRFTOKEN'
 
     export default Vue.extend({
         data: function() {
             return {
-                errors: [],
-                hasErrors: false,
                 questionnaire: {},
             }
         },
@@ -51,40 +40,11 @@
         },
         methods: {
             back: function() {
-                this.clearErrors()
                 this.$emit('back')
             },
-            clearErrors() {
-                this.errors = []
-                this.hasErrors = false
-            },
             done: function() {
-                if (this.questionnaire.end_date) {
-                    this.questionnaire.end_date = moment(String(this.questionnaire.end_date)).format('YYYY-MM-DD')
-                } else {
-                    delete this.questionnaire.end_date  // remove empty strings, it throws date format error.
-                }
-
-                console.log('Questionnaire to save : ', this.questionnaire)
-                this.saveQuestionnaire()
+                this.$emit('save-questionnaire')
             },
-            saveQuestionnaire(){
-                this.clearErrors()
-                let saveMethod = axios.post.bind(this, save_questionnaire_url)
-                if (typeof this.questionnaire.id !== undefined) {
-                    this.questionnaire.is_draft = false
-                    saveMethod = axios.put.bind(this, save_questionnaire_url + this.questionnaire.id + '/')
-                }
-                saveMethod(this.questionnaire)
-                    .then(response => {
-                        console.log(response)
-                        window.location.href = '/accueil/'
-                    }).catch(error => {
-                        console.log(error)
-                        this.hasErrors = true
-                        this.errors = error.response.data
-                    });
-            }
         },
         components: {
             QuestionnaireDetail
