+++ templates/ecc/questionnaire_create.html
@@ -1,10 +1,6 @@
 {% extends "base.html" %}
 {% load static %}
 
-{% block page_title %}
-Nouveau questionnaire
-{% endblock page_title %}
-
 {% block extra_static_header %}
 {{ block.super }}
 {% endblock extra_static_header %}
